refactor: 시스템 클래스와 유효성 검사 클래스 분리
