+++ src/main/java/subway/Application.java
@@ -1,12 +1,14 @@
 package subway;
 
 import java.util.Scanner;
+import subway.manager.SubwayManager;
+import subway.system.SubwaySystem;
 
 public class Application {
     public static void main(String[] args) {
         final Scanner scanner = new Scanner(System.in);
         // TODO: 프로그램 구현
-        SubwayManager subwayManager = new SubwayManager();
-        subwayManager.manage(scanner);
+        SubwaySystem subwaySystem = new SubwaySystem();
+        subwaySystem.run(scanner);
     }
 }
