+++ src/main/java/subway/manager/LineManager.java
@@ -1,81 +1,25 @@
-package subway;
+package subway.manager;
 
 import java.util.Scanner;
+import subway.validator.LineManagerValidator;
 import view.LineManagerView;
 
 public class LineManager {
 
-    public static final String ASK_OPTION_MESSAGE = "## 원하는 기능을 선택하세요.";
     public static final String ERROR_PREFIX = "[ERROR] ";
-    public static final String WANT_QUIT_CODE = "B";
-    public static final String OPTION_ERROR_MESSAGE = "1~3 또는 B 옵션 중 하나를 입력하세요";
-    private static final int OPTION_MIN = 1;
-    private static final int OPTION_MAX = 3;
     private static final int MIN_NAME_LENGTH = 2;
     public static final String INPUT_LINE_MESSAGE = "## 등록할 노선 이름을 입력하세요.";
     public static final String INPUT_UPWARD_MESSAGE = "## 등록할 노선의 상행 종점역 이름을 입력하세요.";
     public static final String INPUT_DOWNWARD_MESSAGE = "## 등록할 노선의 하행 종점역 이름을 입력하세요.";
-    public static final String LINE_NAME_INPUT_ERROR_MESSAGE = "존재하지 않는 2자 이상의 노선 이름을 입력하세요";
     public static final String LINE_NAME_DELETION_ERROR_MESSAGE = "존재하는 2자 이상의 노선 이름을 입력하세요";
-    public static final String STATION_NAME_ERROR_MESSAGE = "존재하는 2자 이상의 역 이름을 입력하세요";
     public static final String INFO_PREFIX = "[INFO] ";
     public static final String ENROLLMENT_LINE_INFO_MESSAGE = "지하철 노선이 등록되었습니다";
     public static final String ASK_DELETION_MESSAGE = "## 삭제할 노선 이름을 입력하세요.";
     public static final String DELETION_INFO_MESSAGE = "지하철 노선이 삭제되었습니다";
     public static final String DELETION_FAIL_MESSAGE = "지하철 노선이 존재하지 않거나 삭제할 수 없습니다";
-    public static String userOption = "";
 
-    public static void manage(Scanner scanner) {
-        LineManagerView.printLineManagerMainScreen();
-        userOption = getUserOption(scanner);
-        callOptionMenu(userOption, scanner);
-    }
 
-    public static String getUserOption(Scanner scanner) {
-        try {
-            System.out.println(ASK_OPTION_MESSAGE);
-            userOption = scanner.nextLine();
-            System.out.println();
-            validateUserOption(userOption);
-            return userOption;
-        } catch (IllegalArgumentException iae) {
-            System.out.println(ERROR_PREFIX + iae.getMessage());
-            return getUserOption(scanner);
-        }
-    }
-
-    public static void validateUserOption(String userOption) throws IllegalArgumentException {
-        if (userOption.equals(WANT_QUIT_CODE)) {
-            return;
-        }
-        int optionNumber;
-        try {
-            optionNumber = Integer.parseInt(userOption);
-            if (optionNumber < OPTION_MIN || optionNumber > OPTION_MAX) {
-                throw new IllegalArgumentException();
-            }
-        } catch (Exception e) {
-            throw new IllegalArgumentException(ERROR_PREFIX + OPTION_ERROR_MESSAGE);
-        }
-    }
-
-    public static void callOptionMenu(String userOption, Scanner scanner) {
-        if (!Character.isDefined(userOption.charAt(0))) {
-            return;
-        }
-        int optionNumber = Integer.parseInt(userOption);
-        if (optionNumber == 1) {
-            enrollLine(scanner);
-        }
-        if (optionNumber == 2) {
-            deleteLine(scanner);
-        }
-        if (optionNumber == 3) {
-            searchLine(scanner);
-        }
-    }
-
-    public static void enrollLine(Scanner scanner) {
+    public void enrollLine(Scanner scanner) {
         System.out.println(INPUT_LINE_MESSAGE);
         String lineName = inputLineNameForEnrollment(scanner);
         System.out.println(INPUT_UPWARD_MESSAGE);
@@ -91,7 +35,7 @@ public class LineManager {
     public static String inputLineNameForEnrollment(Scanner scanner) {
         String lineName = scanner.nextLine();
         try {
-            validateLineNameForEnrollment(lineName);
+            LineManagerValidator.validateLineNameForEnrollment(lineName);
             return lineName;
         } catch (IllegalArgumentException iae) {
             System.out.println(ERROR_PREFIX + iae.getMessage());
@@ -102,7 +46,7 @@ public class LineManager {
     public static String inputStationNameForEnrollment(Scanner scanner) {
         String stationName = scanner.nextLine();
         try {
-            validateStationNameForEnrollment(stationName);
+            LineManagerValidator.validateStationNameForEnrollment(stationName);
             return stationName;
         } catch (IllegalArgumentException iae) {
             System.out.println(ERROR_PREFIX + iae.getMessage());
@@ -110,27 +54,7 @@ public class LineManager {
         }
     }
 
-    public static void validateStationNameForEnrollment(String stationName)
-        throws IllegalArgumentException {
-        if (stationName.length() < MIN_NAME_LENGTH) {
-            throw new IllegalArgumentException(STATION_NAME_ERROR_MESSAGE);
-        }
-        if (!SubwayManager.isDuplicatedStation(stationName)) {
-            throw new IllegalArgumentException(STATION_NAME_ERROR_MESSAGE);
-        }
-    }
-
-    public static void validateLineNameForEnrollment(String lineName)
-        throws IllegalArgumentException {
-        if (lineName.length() < MIN_NAME_LENGTH) {
-            throw new IllegalArgumentException(LINE_NAME_INPUT_ERROR_MESSAGE);
-        }
-        if (SubwayManager.isDuplicatedLine(lineName)) {
-            throw new IllegalArgumentException(LINE_NAME_INPUT_ERROR_MESSAGE);
-        }
-    }
-
-    public static void deleteLine(Scanner scanner) {
+    public void deleteLine(Scanner scanner) {
         System.out.println(ASK_DELETION_MESSAGE);
         String lineName = inputLineNameForDeletion(scanner);
         if (SubwayManager.deleteLine(lineName)) {
